Add FluentValidation

+++ WasteSegregation.Application/Dto/CreateWasteBagsDto.cs
@@ -2,9 +2,9 @@
 
 public class CreateWasteBagsDto
 {
-    public byte? BlueBag { get; set; }
-    public byte? GreenBag { get; set; }
-    public byte? YellowBag { get; set; }
-    public byte? BrownBag { get; set; }
+    public int? BlueBag { get; set; }
+    public int? GreenBag { get; set; }
+    public int? YellowBag { get; set; }
+    public int? BrownBag { get; set; }
     public DateTime? ReceiptDate { get; set; }
 }
