+++ WasteSegregation.Application/Dto/WasteBagDto.cs
@@ -3,9 +3,9 @@
 public class WasteBagDto
 {
     public int Id { get; set; }
-    public byte? BlueBag { get; set; }
-    public byte? GreenBag { get; set; }
-    public byte? YellowBag { get; set; }
-    public byte? BrownBag { get; set; }
+    public int? BlueBag { get; set; }
+    public int? GreenBag { get; set; }
+    public int? YellowBag { get; set; }
+    public int? BrownBag { get; set; }
     public DateTime? ReceiptDate { get; set; }
 }
