+++ WasteSegregation.Domain/Entities/WasteBag.cs
@@ -3,12 +3,11 @@
 public class WasteBag
 {
     public int Id { get; set; }
-    public byte? BlueBag { get; set; }
-    public byte? GreenBag { get; set; }
-    public byte? YellowBag { get; set; }
-    public byte? BrownBag { get; set; }
+    public int? BlueBag { get; set; }
+    public int? GreenBag { get; set; }
+    public int? YellowBag { get; set; }
+    public int? BrownBag { get; set; }
     public DateTime? ReceiptDate { get; set; }
-
     public string UserId { get; set; }
 
     public int RealEstateId { get; set; }
